fix: es6 import paths, remove amd output, remove old RTE configurations, remove old RTE plugin
