+++ Configuration/RTE/IconPicker.yaml
@@ -1,20 +1,17 @@
 editor:
-  externalPlugins:
-    IconPicker: { resource: 'EXT:bw_icons/Resources/Public/JavaScript/Plugins/IconPicker/plugin.js', route: 'ajax_icon_selection' }
   config:
     contentsCss: 'EXT:bw_icons/Resources/Public/Css/Rte.css'
-    extraAllowedContent:
-      - img[!src,data-icon-name,data-icon-base-name,loading]
-      - i[*](*)
-    coreStyles_italic:
-      element: 'i'
-      overrides: 'span'
-    extraPlugins:
-      - IconPicker
+
+    externalPlugins:
+      IconPicker: { route: 'ajax_icon_selection' }
+
     importModules:
       - '@blueways/bw-icons/IconPicker.js'
+
 processing:
   allowTags:
     - i
-  HTMLparser_db:
-    noAttrib: b,u,br,center,hr,sub,sup,strong,em,li,ul,ol,blockquote,strike
+    - img
+  allowTagsOutside:
+    - img
+    - i
