--- Resources/Public/JavaScript/Plugins/IconPicker/lang/de.js
@@ -1,9 +1,0 @@
-CKEDITOR.plugins.setLang(
-    'IconPicker',
-    'en',
-    {
-        toolbar: 'Icon',
-        save: 'einfügen',
-        modalTitle: 'Icon Auswahl'
-    }
-);
