--- Resources/Public/JavaScript/Plugins/IconPicker/lang/en.js
@@ -1,9 +1,0 @@
-CKEDITOR.plugins.setLang(
-    'IconPicker',
-    'en',
-    {
-        toolbar: 'Icon',
-        save: 'Insert icon',
-        modalTitle: 'Select icon'
-    }
-);
