+++ tsconfig.json
@@ -19,7 +19,7 @@
 		"experimentalDecorators": true,
 		"paths": {
 			"@blueways/bw-icons/*": [
-				"Resources/Public/ECMAScript6/*"
+				"Resources/Public/JavaScript/*"
 			],
 			"@typo3/backend/*": [
 				"vendor/typo3/cms-backend/Resources/Public/JavaScript/*"
